pyproject.toml
[tool.poetry]
name = "fastapi-ecommerce-postgresql"
version = "1.0.0"
description = "FastAPI Ecommerce Project with PostgreSQL"
authors = ["Zeeshan Asim <zeeshanravian1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.2"}
fastapi-mail = "^1.4.1"
sqlalchemy = {extras = ["all"], version = "^2.0.21"}
asyncpg = "^0.28.0"
alembic = "^1.12.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"
black = "^23.9.1"
ipykernel = "^6.25.2"
pylint = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
